+++ biblioteca/src/main/java/com/residencia/biblioteca/services/LivroService.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+
 import com.residencia.biblioteca.entities.Livro;
 import com.residencia.biblioteca.repositories.LivroRepository;
 
@@ -21,7 +22,7 @@ public class LivroService {
 		}
 	
 	public Livro buscarLivroPorId(Integer id) {
-		return livroRepo.findById(id).get();
+		return livroRepo.findById(id).orElse(null);
 		
 	}
 	
@@ -34,8 +35,23 @@ public class LivroService {
 		return livroRepo.save(livro);
 	}
 	
-	public void deletarLivro(Livro livro) {
-	livroRepo.delete(livro);
+	public Boolean deletarLivro(Livro livro) {
+		if (livro == null)
+			return false;
+
+		Livro livroExistente = buscarLivroPorId(livro.getCodigoLivro());
+		if (livroExistente == null)
+			return false;
+
+		livroRepo.delete(livro);
+
+		Livro livroContinuaExistindo = buscarLivroPorId(livro.getCodigoLivro());
+		if (livroContinuaExistindo == null)
+			return true;
+
+		return false;
+}
+	
 		
 	}
-}
+
